pyproject.toml
[tool.poetry]
name = "cw-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
ipython = "^8.29.0"
django-cors-headers = "^4.6.0"
requests = "^2.32.3"
drf-yasg = "^1.21.8"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
python-telegram-bot = "^21.6"
eventlet = "^0.37.0"
redis = "^5.2.0"
coverage = "^7.6.4"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
